Update text to clarify spec behavior
The overview mentions that the color space is shared with the
window system but there is no specification text that says that
so adding that in this version.
Generally such information would need to be shared with the window
system by the application and if that is still required this
update doesn't make sense.