EXT_image_dma_buf_import_modifiers: Correct no-modifier-token behaviour
Zero is a valid and explicit modifier token, meaning that the image is
pitch-linear, i.e. no tiling or compression is applied.

When no modifier is passed for dmabuf import, the implementation may
assume the image is linear, but is not required to do so. Many drivers
which supported tiling formats before modifiers (or which do not yet
support modifiers) will determine the layout via other mechanisms, e.g.
metadata held by the kernel.

Fix the wording for the case where the user does not explicitly pass
modifiers to state that the layout will be determined by the
implementation, rather than stating the layout must be linear.

Reported by @olvaffe