Correct and clarify the EGLSync text in EGL_NV_stream_consumer_eglimage
Correct the dependencies of EGL_NV_stream_consumer_eglimage to specify
EGL_EXT_sync_reuse instead of EGL_KHR_reusable_sync.

Rework the description of the <sync> parameter in
eglStreamAcquireImageNV and eglStreamReleaseImageNV to clarify its
behavior, and to specify that it's supposed to be of type
EGL_SYNC_FENCE.

Also add a few missing error codes.