Fix extension/registry mismatch for EXT_surface_compression
EXT_surface_compression defines eglQuerySupportedCompressionRatesEXT
as taking an "EGLConfig config" argument.

However, egl.xml defines it as taking "EGLConfig* configs".

This function is meant to return a list of supported rates for a
"combination of display and EGLConfig", so it seems like the declaration
in egl.xml is the wrong one.