EGL_KHR_platform_wayland: clarify EGLSurface locking
When querying EGLSurface for EGL_KHR_BUFFER_AGE or rendering to it
it will resuilt in locking its back buffer until the next swap.
This behavior is long established in mesa and is intentional.

The rationale behind this behavior is to jarring partial content
that could be if mesa throwed the back buffer in the middle of rendering
due to resize.

This commit explicitly states this behavior, so it won't be
surprising for the end users and also adds a dependency on
EGL_EXT_buffer_age.